# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Finance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /summary/spending:
    get:
      operationId: getSpendingSummary
      tags: [summary]
      summary: Get spending summary by category for a period
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Spending summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"
...
  /summary/recent:
    get:
      operationId: getRecentActivity
      tags: [summary]
      summary: Get recent transactions with enriched data
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        date:
          type: string
        paymentMethod:
          type: string
        userId:
          type: integer
        householdId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - amount
        - description
        - categoryId
        - categoryName
        - categoryColor
        - categoryIcon
        - date
        - paymentMethod
        - userId
        - householdId
        - userName
        - createdAt
    TransactionInput:
      type: object
      required:
        - amount
        - description
        - date
        - paymentMethod
      properties:
        amount:
          type: number
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        date:
          type: string
        paymentMethod:
          type: string
    TransactionUpdate:
      type: object
      properties:
        amount:
          type: number
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        date:
          type: string
        paymentMethod:
          type: string
...
    CategorySpending:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        categoryIcon:
          type: ["string", "null"]
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
      required:
        - categoryId
        - categoryName
        - categoryColor
        - categoryIcon
        - total
        - count
        - percentage
...