# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionUpdate {
  amount?: number;
  description?: string;
  /** @nullable */
  categoryId?: number | null;
  date?: string;
  paymentMethod?: string;
}